kabi.c: fix redundant output from get_mycontainer().
get_mycontainer() was stopping at the first instance of the "me" field,
so it was always presenting the same container name. Fix was to set a
flag in the containermap to indicate that the current instance has
already been read. If we want to go through the list again, we'd have to
reset all these flags first, but we don't do that, so not a problem.

Did some other cleanup. get_container_name function did not need to
traverse the sublist of symbols to get the name field. It was at the
top.