kabi.c - add some new options, fix some old bugs
Code now includes a show-users option that shows the users (containers)
of the nested symbols that have been used elsewhere in the code.

Couple of bugs were exposed by this new code. In get_declist(), string
typnam has to initialized, so initialized it to '\0';

Also needed to add some control to the logic to only add compound types
to the usedlist.