kabi.c: Give each file its own knode for multiple file processing
Design flaw was that the knode at the very top of the tree was the
only one getting a filename. Consequently, the only FILE recorded
and reported was the first one encountered in the list. Fixed by
providing separate knodes for each file encountered.

Debug code is enabled for now.