kabi.c: fix infinite recursions
Only happened in certain files. I had the case where a struct pointed to
itself, but I didn't have the case where the struct pointed to another
struct that then pointed back to the first struct. This patch takes care
of that using the redundancy check, rather than testing for ancestry. As
a beneficial side effect, redundancy is minimized.

Propagate the name of the file through all the descendant knodes. This
information could be helpful in the future, but for now I'll relegate
that to the perl kabilist-search script.

Turned off debugging.