kabilist.sh: submit one file at a time to kabi-parser
At first it seemed that it might be faster to pass all the files to
kabi-parser on the command line at one time. However, there were some
real problems with that, the worst one being that it required way more
memory and processing power than found in the average workstation.

Even on a system with 32G of memory, things started bogging down badly
after about the thousandth file, with still another 18,000 or so to go.

One-file-at-a-time processing only takes about 15 minutes on the average
i7 with 4 cores.