kabi.c: Add option for database file
Added -k option to name the database file in which to store the
kabi tree. Also added a test to see if the file already exists.
Previously, we just blew a hole in it and started from scratch,
but you can't do that when you're processing more than one file.

TODO:
. Need to move the creation of the virtual table in the database
  from kabi.c into kabilist.sh. That's something you only want to
  do once, after the whole database has been created.

. Not getting nested data types from the RETURN objects, so that
  must be investigated.