kabilookup.c: Got the memory resident table working
Needed this to be able to print the results of the search in reverse
order. There is a way to order the results of a search from sql, but I
haven't found a reliable way to connect id == parentid to a sql select
statement (database search).

However, the database is fast and useful for retrieving the first level
searches.