kabilookup.c: add -x option for exports only
Also cleaned up the parsing logic and added command line error parsing.
Enhanced the help message to introduce the new option

TODO:
The option does not seem to be working yet. It appears that the sql
statement may be too long for the sqlite_mprintf() function to handle.
May have to break the statement up and change it from unions to joins.