kabilookup.c: --no-dups is working for non-verbose struct search.
Only unique exported symbols are presented, even though a given struct
may appear many times within the nest of an exported symbol.

TODO:
Need to find a way to skip rows that are repeats. Need to find a way to
locate the primordial ancestor before printing the symbol and its outer
(containing) nest. I think this is an application for an array of rows
struct. There's probably a way to do it using the database sets.