kabi.c: moved output formatting block and started database enhancements
Rather than have one redlist of already-seen symbols for the whole file,
maintain a separate list for each EXPORTED function within the file.
This assures that we will find a given symbol for every exported function
in the file affected by it.

Previous scheme at minimizing duplication was too aggressive, so we did
not always find every exported function in a file for a given symbol.