kabi.c: put the redlist in each EXPORTED symbol's knode
Having one redlist of previously encountered symbols for each file was
problematic, because symbols that were used by other exported symbols in
the same file would not be reported anew. A lookup of a symbol would
then only return the exported symbols in a file that were first
encountered having that symbol, while other exported symbols in that
file that used the symbol would not be reported in the lookup.

The reason for the policy of only recording a symbol once is to minimize
the size of the database and the subsequent reporting of a lookup.
Unfortunately, the once-per-file symbol policy was too agressive. This
changes the policy to once-per-exported-symbol.

The database will be considerably larger, but that's the cost of
accuracy, in this case.