kabilookup.c: bug fixes resulting from detailed testing
This culminates a series of fixes starting with commit 3fd6fcdc.

There were inconsistencies in the reporting of verbose and concise
nested compound types.

Changes had to be made to the kabi-parser (kabi.c) in order to obtain
a more accurate picture of which exported symbols were affected by which
nested compound data types. Having a less aggressive redundancy filter
fixed that problem, but presented new challenges to the kabi-lookup.

Unfortunately, the less aggressive redundancy policy made the
kernel-wide database 10 times larger, which introduces performance
problems that will have to be addressed in a different manner.