Made changes to everything in order to move the code to a graph model
Implementing boost serialization. May eventually move to a boost graph
model, but since this database is read-only once it's built, there is
probably no need for a dbms. We shall see.

The reason for all these changes is to reduce the size of the data base
and increase the access speeds.

This code does not compile yet, but I am committing it in order to
preserve it for now. It's currently on the devel branch, so it will not
affect the master branch if anyone wants to clone that and use it.