Preparation for using boost to serialize the data.
Can't use containers of pointers for boost. Boost can dereference
pointers, but not pointers to pointers. Modified the kabi-node.*
accordingly.

This commit is a good stopping point before removing the knode code
from the kabi-parser, since we won't need it anymore.

This code compiles and executes as expected, but it's far from done.