Got the knode out of the sparse mining code
This code builds and executes.

Now to replace write_knodes with calls to the serialization code to
serialize the qnodes. Don't know yet whether to serialize the knodes and
cnodes together or separately. The problem is that, once the whole
database is built, there has to be a cleanup stage that goes through the
whole data base and compresses the code.

Each file in the kernel generates a list of symbols. The symbols are not
duplicated for the file, and a list of parents and children is
maintained for each symbol in the qnodes. After all the files in the
kernel are processed, there will be duplicate symbols in the database,
because the code does not look at the contents of the file for
duplicates. This would take too much time. It's faster to post process
the duplicates all at once.

Post processing will consist of deleting duplicates and attaching their
parent and children lists to the one and only symbol. This will compress
the database considerably, especially since the parent and children
lists consist of structs having only two integers.