kabi-data.sh kabi.c: added -c switch for cumulative database
Don't know if this is the best solution, but it works. The database is
read into a vector with each invocation of the kabi-parser. The database
file is destroyed and then rewritten with the old serialized data and
the new. This does compress the database considerably, but it takes
almost an hour to do it.

Changed the options for database filenames from -b to -f.