kabi-node.cpp: update the sublist when updating the qnodelist
The sublist is a subset of the data in the qnodelist. We are only
interested in duplicates of compound data types, so no need to include
all the basetypes in our searches. This change is in hopes of improving
performance when building the database cumulatively from multiple calls
to the kabi-parser.