kabi-compdb: complete support for external database compression
This utility is intended to be called after calling kabi-parser on
multiple files while sending its output to the same database file.

Boost serialization will append a new serialized list to the end of the
database file for every .i file processed.

The first problem is that boost will only deserialize the first list in
the datafile, because the number of records is the first thing it reads.
Each serialization from each .i file will have its own number of
records.

This utility reads each list in the datafile into one big list in memory
and then serializes this new list to the original database file.

The second problem is that kabi-parser only sorts duplicates for the
file it's currently processing. That means that there are duplicates
among the multiple files.

This utility filters out those duplicates while creating the new big
list in memory.