kabi-map, kabi.c: now using std::multimap for node storage
The hope is that the red-black btree sort and search algorithms in the
std::multimap container will be a big enough improvement on compressing
the database to make a real difference.

This code builds cleanly, but it remains to be seen if it executes. That
will have to wait until tomorrow.