kabi.c, kabi-map: changed some func names and added a trimmer func
Changed extract_type to get_decl, and add_to_declist to add_to_decl.
The original names were legacies from the time that kabi.c contained
all these routines.

Also added trim_decl(). When we build the decl (delcaration list),
we add a space to the end of each word. We want to remove that space
before calculating the crc for the datatype part of the declaration.
The full declaration would include the name (identifier), but we are
only interested in the crc for the data type, as in "struct device".