kabi-map.cpp: breakthrough on the duplicates front
The database graph edges are organized as an std::multimap, with the crc
of compound data structure types used as a key for them, and fully
qualified identifiers, type and name, used for all other symbols. The
reason for using a multimap is that it is possible for there to be
multiple instances of the same compound data type. Functions always have
unique identifiers for any given namespace, and exported functions must
be unique kernel wide.

We want to capture the multiple instances of compound data types, but we
don't want to have children and parents maps scattered among the several
instances. When that happens, it is difficult, if not impossible, to
determine which child belongs to which instance.

We also do not want many multiples of the same instances at the same
level of the hierarchy, as the reporting of such would be burdensomely
redundant.

The method for determining the correct primordial parent is illustrated
by the following pseudo code representation of the new function
qn_lookup_parent().

1. IF we have never encountered an instance of this compound type before
   THEN INSERT the new instance in the map.

2. IF there is already one instance, but only one, of the compound type
       we are currently processing,
   THEN update the one instance before inserting the new instance.

3. IF there is more than one instance of the compound type we are
       currently processing,
   THEN the primordial one is identified as the only one having a children
       map of size greater than zero, because of state 2 above.

TODO:
Need to be able to discern between new instances and child instances.
CTL_BACKPTR flag will help in that respect. Instances that occur in
different hierarchies will have to be separated and identified.

Also added functions to insert pairs into the map. If I want to change
the insert with emplace in the future, I will only have to do in this
one place.