Changing from monolithic database to distributed
This is a performance enhancer all around. Rather than having to
compress everything into one big data file and then do lookups in a
large file, this model distributes the database in pieces. Every file
that's run through the parser has its own database file in its own
directory. These can be cleaned up easily with the find utility, or they
can be located in a directory that's not followed by git.

Performance with this method is orders of magnitude better than anything
I've yet been able to achieve with the monolithic approach.

	modified:   kabi-data.sh
	modified:   kabilookup.cpp
	modified:   kabilookup.h