kabilookup: finish migration to distributed database model
Changed help to inform user that -f switch selects a different list of
files than the default built by kabi-data.sh

Changed the error message for failure to open the filelist.

Created a std::pair for the error index and datafile name. A pair
will be stored in a class member vector for every error related to
a search string and a data file. This gives a list of files in which
the search string could not be found. These error messages are then
directed to cerr, not cout.

Changed the exe_count() to use "\r" to update the count for each file
processed and cout << endl after the last count.

Added a default for the m_filelist,
      "./redhat/kabi/parser/kabi-files.list"