kabi-map: refine qn_is_dup()
This function was returning true if the compound type was encountered
anywhere in the file. This change looks for dups at the ARG level,
rather than at the FILE level. We get a lot more data, and more
accuracy, but we lose a bit of performance and consume more disk space.
It will be up to the users to determine if this much depth is necessary.

Of course, I can always make it an option. --dup-level, so that's a
thought.