least-compression
The code at this state is being tagged least-compression. This has the
benefit of providing the most information, but it does compromise
lookup performance by up to 50% and takes about 5 times as much disk
space. We will branch from here with more compression to see if we can
still glean the same amount of data and enjoy the higher performance.

The way compression is being done is by limiting the reporting of
duplicate symbols. This is done in function kabi-map.cpp::qn_is_dup().
The policy for reporting duplicates is implimented there.

The kabi-map.h file has an explanation of what constitues and
differentiates the parent, ancestor, and function cnodes.

A cnode is simply a std::pair consisting of the crc of the symbol's
declaration and the level in the exported symbol's hierarchy at which
the symbol appears.

The parent crc of a qnode is the crc of the compound type in which the
symbol is instantiated as a member.

The ancestor crc of a symbol is the crc of the compound typed argument
of an exported function from which the symbol was descended.

Currently, the duplicate logic only reports as duplicates those
compound-typed symbols, that is, symbols that are characterized by
sparse as having their own symbol lists, that have already been
encountered in the file being processed AND that whose ancestor has
already been reported.

This means that when processing moves on to a new argument for the
exported function's symbol, the first compound typed symbol that was
previously reported as a duplicate, will not be flagged a duplicate.
However, all subsequent duplicates having the same ancestor will be
flagged as such.

This allows for the first instance of a symbol under an argument of an
exported function to have its members fully explored and reported, even
though that has already been done for the same symbol under a previous
argument or function in the same file.

This is the least compression that still allows for good reporting and
acceptable performance.