least-compression
The code at this state is being tagged least-compression. This has the
benefit of providing the most information, but it does compromise
lookup performance by up to 50% and takes about 5 times as much disk
space. We will branch from here with more compression to see if we can
still glean the same amount of data and enjoy the higher performance.