kabiparser: trying to increase compression without losing information
This branch increases compression by reporting and and all duplicate
compound types in a file as dups, so that we do not traverse the the
symbol list for any duplicates. Rather, what this tries to do is copy
the children list from a previous compound data type having the same
crc.

Sometimes clever is the enemy of smart. Haven't gotten this working
successfully yet, because the level at which the other children appear
is different than the level at which the duplicate's children will
appear. This presents a dead-end to lookup. There may be a way to
address this, but I haven't thought of one yet that isn't more
complicated than simply reducing the compression.