kabi-map: creating parents map to replace single parent field in qnode
This commit does not compile yet. Hope to get that happening tomorrow.

This is in hopes of higher compression and higher accuracy. The idea is
to have a map of parents, rather than one parent. With one parent, I
would need a qnode for every instance of the corresponding symbol at
every level. With a parents list, I can simply insert cnodes, which are
much smaller, and have just one qnode with a parents map and children
map.

This will also yield more accuracy, since every instance of the symbol
can be captured at every level in the cnode. I've tried this before, but
this time, instead of just the crc of the declaration and the level, the
cnode also contains the crc of the exported symbol at the top. This
gives it a frame of reference the previous implemenation did not have.

With the function crc, we can accurately trace back the qnode from every
instance at every level to its correct source at the exported function.

We are not tracing back to the argument, because we really only want to
have a full tree exposed only once for each function. Exposing it for
every argument would be too noisy. Also, it is possible to have more
than one argument with the same crc signiture, as in ...

doit("struct foo *one", struct foo *two)

Both args would have the same signature, and their descendents would be
completely expanded, when they should simply be flagged as dups. One
descendency tree is plenty enough.

Of course, this can always be changed, if desired. Again, perhaps with a
possible --dup-level switch.