kabi-map.cpp: Change qn_is_dup policy to increase compression
I think this is probably the best compromise. Not every function in a
file will show the entire struct tree to its very depths. However,
compound data types will be discovered under every arg and return of
every exoported function.

It takes a minute less to build this slightly more compressed database
than commit 0324592b, and lookup times are improved as well.

Commit 0324592b didn't quite work as advertised, because args of
different exported functions in the same file frequently shared the same
crc signature. The correct way to do what I was trying to do there was
to scope the duplicate detection at the function level. This generated a
much larger database, with attendant performance problems.

We can always choose this as an option in the future.

TODO:
It might be nice to restrict lookups to the directory level. This can be
done without too much trouble, so I will mention it to my commissioners.