kabi-map: Intermediate commit
Separating qnode into qnode and dnode. qnode interfaces with the kabi.c
parser. dnode interfaces with th c++ side. qnode is a struct, dnode is a
class.

qnode: vehicle to pass info to the c++ side in kabi-map

dnode: contains the declaration of a specific data type

cnode: contains the identifier and hierarchical level at which the data
       type was discovered.