kabi-data.sh: update to current reality
Updated the instructions and accompanying commentary. We no longer build
a monolithic database, and we no longer use SQLite. No longer have an
optional error file, since we are working on quieting sparse by creating
options that forego semantic parsing.

Also fixed the directory variables so that the complete path appears in
the file-list. This makes it possible to do searches from any directory,
rather than only from the top of the linux tree.