kabi-map: some refactoring and comment updating
Compiles and runs.

TODO:
Get the debug dump to correctly represent the contents of the data file.

Added a template insert_node function for times we don't care about the
iterator telling where the item was inserted in the map. When we are
using the pointer to the item AFTER inserting it in the map, we need to
have the value of the iterator (pointer) to where it was stored, so that
we are updating the correct item.

in kb_is_dup, cannot directly return whether the iterator was pointing
beyond the end of the map. Must first assign to a local bool and return
that. This seems like a compiler bug, but we can work around it.