kabilookup.cpp: got the whole word side of -d declaration working
This simply dumps a data structure and its children. In verbose mode
it dumps the children's children, too.

The inexact word match seems to run off the stack in recursion land.
This is the last part of lookup that needs to be working to put it
in an alpha state.