rownan.cpp: reset normalized parameters after each dump
Row Manager was not resetting the normalized level and normalized
boolean after being invoked to perform a normalized dump. As a
consequence, negative row levels were ocurring, which caused the
indent() routine to wind itself off the stack.

A normalized dump sets the first row level to zero, regardless of the
indentation level of the first row, in order to present a "normalized"
view of the data in the rows. Used in kabilookup to view a struct
outside of its context as used in the code.