kabilookup.cpp: fix error reporting logic
Using -u subdir option, error logic was reporting "not found", even
though we had found the search string many times in many files. The last
file parsed did not have the string, so the "not found" message was
reported. The fix was to check to see if the string was ever reported
and to change the status to EXE_OK if it was ever found anywhere.