kabi-data.sh: need to grep for ksymtab after all
Thought it would be enough for makei.sh to grep for EXPORT_SYMBOL.
However, there is conditionally compiled code where exported symbols
might not be included in the compile. So, we will have a .i file,
but the kabi-parser will not generate a graph for it, because there
are no ksymtab elements.

Also commented out the line that looks for kabi-parser in the
redhat/kabi directory, since the rpm install is locating it in
/usr/sbin for now.