kabi-map.cpp: always insert child into parent dnode children map
In kb_update_nodes() ...
The call to insert the current dnode's cnode into its parent's children
map was not happening if the current dnode had been seen before, that
is, identified as a dup by kb_is_dup(). It was an error in logic that
the insert was not done if it was a dup. We only want to exclude
inserting the dnode into the global dnode map, since there's already one
there.