kabi-map.cpp: kb_init_crc was acting on parent flags
This bug corrupted the parent lookup at the ARG level, because the crc
for the "argument" part of the ancestry was wrong, being based upon the
parent's flags.

Also added a debug statement. Decided to leave it in, as it is inactive
when NDEBUG is defined, which is the default.