kabilookup.cpp: rethink the parent lookup
The previous method involved an elaborate scheme to minimize the
repetition displayed when a given struct could be found in more than one
place in the same tree under a function/arg.

This method is more noisy, but also more accurate and more maintainable.
We can revisit this if the users find it too noisy.