kabilookup: fixed 2 bugs and did some refactoring
First bug was occasional seg fault spinnig off the stack. In
get_children(), the test for CTL_BACKPTR was insufficient. Adding
a test for CTL_ISDUP fixed the bug.

Second bug was that the filename did not get printed on lookup of
exported functions. Fix was to add a small routine that finds the
parent of the exported function, which happens to be its file.

Some minor refactoring and comment headings for functions that
didn't already have them.