kabilookup: Update help text and change cout to cerr as appropriate
Send info to cerr while traversing the files, i.e, filenames as they are
being processed. This allows for using the utility in a script without
filling the data target with useless information. The output of the
filenames during the scan is only eye candy for the user, so they can
see that something is actually happening, but it's cruft when run
inside a script.

Change the -u <subdir> option to the more accurate -m <maskstr>
option.