kabi-data.sh: refactoring and logic improvements
Decided to change the name and location of the default list of kb_dat
files from redhat/kabi/parser/kabi-files.list to
redhat/kabi/kabi-datafiles.list. These changes will be made in the
kabi-lookup routine as well.

The change was due in part because we want to include the kabi data
files in the build, which means we must also provide a clean target
for them. We don't have a directory assigned to us yet, so we will
use the current redhat/kabi directory to store the file list.

Changed logic so that this script can now be run from any directory.

Removed the grep for "ksymtab" because the parser code does not create a
kb_dat file unless it finds ksymtab. This increases processing speed by
almost 8%.