kabi-data.sh: add version and fix find -exec logic
The find -exec sh -c was writing the kb_dat filename to the
kabi-datafiles.list, even if the kabi-parser did not create a
database, because there was no ksymtab in the file. To correct this,
first check that the kb_dat file exists before writing it into the
kabi-datafiles.list.

Some minor refactoring and whitespace fixing.