kabi-lookup: fixes and enhancements
When looking for exported symbols, only cnodes with the CTL_EXPORTED
flag set should be considered when traversing the exported symbol's
cnode map. This change was implemented with a new function,
get_siblings_exported(). The exe_exports() function tests for a valid
return value from this new function before going any further.

The get_children() function was changed to add the parent cnode to the
argument list. Previously, we indexed into the child's sibling map using
the order field. Now, we search for the level and ancestry that
corresponds to that of the parent.

A cnode can have the CTL_ISDUP flag set by the kabi-parser when creating
the graph. This flag is set to prevent spinning off the stack as
kabi-parser recursively pursues subordinate symbol lists. This flag is
not necessarily correct at kabi-lookup time, since the lookup does not
necessarily occur in the same order as the original parse. The
kabilookup::is_dup() and member vector m_dups were created to detect
dups in the lookup context.

Minor corrections were made to the help text.

References to no functions that no longer exist were removed.