kabi-map: fixes and refactoring
The cnode= operator was missing two of the fields. Got them all, now.

Replaced static is_child() with more thorough and detailed global
kb_is_adjacent().

Refactored kb_update_nodes() with more detailed commentary. These
comments are important, if the code is going to be maintained. Alos
corrected the content of the cnode's sibling field to point back to
the order/crc pair of the first sibling in the sibling cnodemap,
which is also the first instance of this symbol or declaration in
the file.

Improved the comment in dump_dnmap() to correctly state what is
output by the dump.