kabilookup: refactor and fix minor bugs
Moved adjacency test into global function in kabi-map, kb_is_adjacent().

Removed get_child_sibling() as unnecessary.

Fixed a bug where we occasionally got incorrect indentation on the
output. Backpointers and dups are "virtualized" in a single cnode for
all instances. That means that the level field, which is used for
indentation, is only correct for the first instance. To adjust for this,
the answer was to simply add 1 to the parent cnode's level.