kabi-parser: refactor for GENKSYMS instead of ksymtab
Earlier versions of kabi.c parsed the sparse generated symbol table
for leading __ksymtab_ strings to identify exported symbols. However,
we want to build the .i files with __GENKSYMS__ defined, which does
not generate the leading __ksymtab_ string. These changes traverse
the file to see if the symbol in the sparse generated symbol table
is on an EXPORT_SYMBOL line.

This may be a temporary fix, depending on the performance. The best
way to do this would be to train sparse to create a list of exported
symbols, or to at least identify those lines it did not parse, such
as the unexpanded EXPORT_SYMBOL lines. Traversing a list in memory
will be vastly faster than traversing and retraversing a file.

Some other refactoring was done to clean up the code.

Files changed:
	kabi.c
	kabi-map.cpp