kabi-parser: make the exported symbol prefix an option
Allows us to use cpp .i files compiled with and without __GENKSYMS__
defined. In the former case, the EXPORT_SYMBOL macros are not expanded.
In the latter case, they are, and the exported symbols have a leading
"__ksymtab_ string prefixed to them. The latter case is the default.

Added -p option with two possible values, "gen" and "tabe", where the
former is with __GENKSYMS__ defined and the latter is without.