kabi-lookup: use chdir to execute from top of kernel tree
If kabi-lookup is executed from anywhere other than the top of the
kernel tree, let the user supply a path to the top of the kernel
tree with the -p <path> option. Some refactoring was necessary as
was adding new functions to deal with getting current directory and
changing the directory from which we're executing.