Makefile: added static libsparse and sparse headers
The sparse library is not a rhel or fedora package. It must be dowloaded
and built separately. Rather than put that onus on the kabtools user,
it's best to simply include the libsparse.a static library and the
headers necessary to build kabi-parser using libsparse.a.

A local include directory has been created in the source tree to contain
the necessary sparse headers. This directory and its contents, as well
as libsparse.a, and will be uploaded to the git repo with the other
sources.