Makefile: move the libsparse.a static lib
Because we're building the various arches in a loop, the libsparse.a for
each arch will be copied into /kabi/work/lib to be linked.

The libsparse.a was created for each arch and located at
/work/sparse/redhat/rpm/RPMS/<arch>

libsparse.a is extracted into /work/kabi/lib from its corresponding rpm
by the kabitools Makefile using rpm2cpio and cpio

If I get around to it, I'll make these directories relative or write a
shell that reproduces this build environemnt.