Makefile: removed dependency on uname for distro and arch
Check to see that boost-serialization is installed. Print an error
message and exit, if not.

Then used the distro and arch information from the boost-serialization
rpm to create the DISTRO and ARCH for the build.

Problem with using uname is that the kernel could be called almost
anything, with no guarantee that the distro and arch would be in the
name, or that they would always be in the same place.

Use @echo in the make target for cleaner output.

Use find to clean up the rpm files when starting from scratch.