kabi-graph: hide kabi graph files from git
Append kbg extension and the kabi-datafiles.list to .git/info/exclude.
When deleting kabi graph, also deletes the lines that were written into
.git/info/exclude.

Made clean the default. Removed the -c option and added the -n option
to disable clean.

Added a target to the command line to pass to the kernel make, to
enable incremental builds.